06-personalized-avatar-generator/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
Pillow>=9.5.0
numpy>=2.3.0
requests>=2.31.0
aiofiles>=23.0.0
psutil>=5.9.0
streamlit>=1.36.0
plotly>=5.17.0
pandas>=2.2.0
torch>=2.0.1
diffusers>=0.21.4
transformers>=4.35.2
opencv-python>=4.8.0
scikit-learn>=1.3.0
accelerate>=0.20.0 